+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 - AIDER_DESK_DATA_DIR now properly respected
 - added support for ARGUMENTS placeholder in custom commands
 - reloading project keeps the current session
+- improved performance of model library by using virtualized table for models
+- added support for provider overrides on model level
+- added support for Z.AI Coding Plan provider
+- updated Electron version to fix issues with latest MacOS version
 
 ## [0.33.0]